Experiments with use Network Part #3

+++ WeatherApp/Models/Wind.swift
@@ -5,7 +5,7 @@
 import Foundation
 struct Wind: Codable {
 	let speed: Double?
-	let deg: Int?
+	let deg: Double?
 
 	enum CodingKeys: String, CodingKey {
 
@@ -16,7 +16,6 @@ struct Wind: Codable {
 	init(from decoder: Decoder) throws {
 		let values = try decoder.container(keyedBy: CodingKeys.self)
 		speed = try values.decodeIfPresent(Double.self, forKey: .speed)
-		deg = try values.decodeIfPresent(Int.self, forKey: .deg)
+		deg = try values.decodeIfPresent(Double.self, forKey: .deg)
 	}
-
 }
